migrate/streamlit_app/requirements.txt
neo4j==5.17.0
python-dotenv==1.0.1
pyvis==0.3.2
streamlit==1.31.1
llama-index==0.10.11
llama-index-agent-openai==0.1.5
llama-index-core==0.10.12
llama-index-embeddings-openai==0.1.5
llama-index-graph-stores-nebula==0.1.2
llama-index-graph-stores-neo4j==0.1.1
llama-index-legacy==0.9.48
llama-index-llms-openai==0.1.5
llama-index-multi-modal-llms-openai==0.1.3
llama-index-program-openai==0.1.3
llama-index-question-gen-openai==0.1.2
llama-index-readers-file==0.1.4
langchain==0.1.12
langchain-openai==0.0.8
llama-index-llms-perplexity==0.1.3
pandas